fix(wasm): stop losing reads, closures, datagrams and error codes
Four defects found in review of the poll conversion, all of them cases
where a browser primitive does not behave the way the first cut assumed.

A dropped session clone swallowed a stream. A browser `read()` cannot be
cancelled: once issued, the browser hands the next value to *that*
request. Dropping the handle that made it stranded whatever arrived, so a
surviving clone never learned the stream existed. Reads orphaned by a
dropped handle now go to a shared queue that the next poll adopts before
issuing a fresh read, oldest first, which also keeps the order the browser
fulfills them in.

`RecvStream::poll_closed` could hang. It deferred while bytes sat in our
buffer, on the theory that the caller would drain them and wake us. But it
borrows `&mut self`, so a caller waiting there is a caller who cannot
read, and `stop()` followed by `closed()` wedged forever -- as did a peer
reset arriving behind buffered bytes. The transport's closure is now
reported when it happens; the buffered bytes are still delivered by later
reads, which is also how quinn behaves.

`try_send_datagram` queued without bound. It wrote on every call and
replaced the previous unresolved write, and a browser write cannot be
cancelled, so a media loop on a stalled connection would pile up a backlog
and eventually send it stale. It now sheds when a write is still
outstanding or `desiredSize` reports no room, and returns whether the
browser took it. The poll path keeps waiting instead of shedding; draining
the previous write is what bounds it.

Stream and session error codes above 255 were corrupted. `web-sys` types
`streamErrorCode` as a byte, from a WebIDL draft that has since widened it
to `unsigned long`: its setter sent a larger code as 0 and its getter
truncated one into a different code's meaning. Both directions now go
around the binding via `Reflect`, and the codes stay `u32` out through
`Error::code` and `closed`.

That last change removes the reason `web-transport` truncated its own
`closed` codes to `u8` -- the comments there cite the WASM signature as
the cause -- so both platforms now report the full `u32`.

+++ rs/web-transport/src/quinn.rs
@@ -247,12 +247,11 @@ impl SendStream {
 
     /// Block until the stream is closed by either side.
     ///
-    /// This returns a (potentially truncated) u8 because that's what the WASM implementation returns.
     // TODO this should be &self but requires modifying quinn.
-    pub async fn closed(&mut self) -> Result<Option<u8>, Error> {
+    pub async fn closed(&mut self) -> Result<Option<u32>, Error> {
         match self.inner.stopped().await {
             Ok(None) => Ok(None),
-            Ok(Some(code)) => Ok(Some(code as u8)),
+            Ok(Some(code)) => Ok(Some(code)),
             Err(e) => Err(Error::Session(e)),
         }
     }
@@ -306,14 +305,10 @@ impl RecvStream {
     }
 
     /// Block until the stream has been closed and return the error code, if any.
-    ///
-    /// This returns a (potentially truncated) u8 because that's what the WASM implementation returns.
-    /// web-transport-quinn returns a u32 because that's what the specification says.
-    // TODO Validate the correct behavior.
-    pub async fn closed(&mut self) -> Result<Option<u8>, Error> {
+    pub async fn closed(&mut self) -> Result<Option<u32>, Error> {
         match self.inner.received_reset().await {
             Ok(None) => Ok(None),
-            Ok(Some(code)) => Ok(Some(code as u8)),
+            Ok(Some(code)) => Ok(Some(code)),
             Err(e) => Err(Error::Session(e)),
         }
     }
